pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mfcs"
version = "0.1.0"
description = "A Python library for function calling in LLMs"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "shideqin", email = "shisdq@gmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "openai>=1.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=0.19.0",
]

[project.urls]
"Homepage" = "https://github.com/mfcsorg/mfcs-python"
"Bug Tracker" = "https://github.com/mfcsorg/mfcs-python/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mfcs"] 